refactor: implement custom useDictionaryLocation hook
